+++ apps/explorer/lib/explorer/chain/token/instance.ex
@@ -89,4 +89,8 @@ defmodule Explorer.Chain.Token.Instance do
       select: to_address
     )
   end
+
+  @spec token_instance_query(non_neg_integer(), Hash.Address.t()) :: Ecto.Query.t()
+  def token_instance_query(token_id, token_contract_address),
+    do: from(i in Instance, where: i.token_contract_address_hash == ^token_contract_address and i.token_id == ^token_id)
 end