+++ apps/indexer/lib/indexer/fetcher/sanitize/token_instance.ex
@@ -6,17 +6,15 @@ defmodule Indexer.Fetcher.Sanitize.TokenInstance do
   use Indexer.Fetcher, restart: :permanent
   use Spandex.Decorators
 
-  require Logger
+  import Indexer.Fetcher.TokenInstance
 
   alias Explorer.Chain
-  alias Explorer.Token.InstanceMetadataRetriever
   alias Indexer.BufferedTask
 
   @behaviour BufferedTask
 
   @default_max_batch_size 1
   @default_max_concurrency 10
-
   @doc false
   def child_spec([init_options, gen_server_options]) do
     {state, mergeable_init_options} = Keyword.pop(init_options, :json_rpc_named_arguments)
@@ -38,7 +36,7 @@ defmodule Indexer.Fetcher.Sanitize.TokenInstance do
   @impl BufferedTask
   def init(initial_acc, reducer, _) do
     {:ok, acc} =
-      Chain.stream_token_instances_with_error(initial_acc, fn data, acc ->
+      Chain.stream_unfetched_token_instances(initial_acc, fn data, acc ->
         reducer.(data, acc)
       end)
 
@@ -47,51 +45,19 @@ defmodule Indexer.Fetcher.Sanitize.TokenInstance do
 
   @impl BufferedTask
   def run([%{contract_address_hash: hash, token_id: token_id}], _json_rpc_named_arguments) do
-    fetch_instance(hash, token_id)
+    if not Chain.token_instance_exists?(token_id, hash) do
+      fetch_instance(hash, token_id, false)
+    end
 
     :ok
   end
 
-  defp fetch_instance(token_contract_address_hash, token_id) do
-    case InstanceMetadataRetriever.fetch_metadata(to_string(token_contract_address_hash), Decimal.to_integer(token_id)) do
-      {:ok, %{metadata: metadata}} ->
-        params = %{
-          token_id: token_id,
-          token_contract_address_hash: token_contract_address_hash,
-          metadata: metadata,
-          error: nil
-        }
-
-        {:ok, _result} = Chain.upsert_token_instance(params)
-
-      {:ok, %{error: error}} ->
-        params = %{
-          token_id: token_id,
-          token_contract_address_hash: token_contract_address_hash,
-          error: error
-        }
-
-        {:ok, _result} = Chain.upsert_token_instance(params)
-
-      result ->
-        Logger.debug(
-          [
-            "failed to fetch token instance metadata for #{inspect({to_string(token_contract_address_hash), Decimal.to_integer(token_id)})}: ",
-            inspect(result)
-          ],
-          fetcher: :token_instances
-        )
-
-        :ok
-    end
-  end
-
   defp defaults do
     [
-      flush_interval: :timer.seconds(3),
+      flush_interval: :infinity,
       max_concurrency: Application.get_env(:indexer, __MODULE__)[:concurrency] || @default_max_concurrency,
-      max_batch_size: Application.get_env(:indexer, __MODULE__)[:batch_size] || @default_max_batch_size,
-      poll: true,
+      max_batch_size: @default_max_batch_size,
+      poll: false,
       task_supervisor: __MODULE__.TaskSupervisor
     ]
   end