+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- [#7286](https://github.com/blockscout/blockscout/pull/7286) - Split token instance fetcher
 - [#7246](https://github.com/blockscout/blockscout/pull/7246) - Fallback JSON RPC option
 
 ### Fixes